+++ src/main/java/selection/ExtendedPointMatch.java
@@ -10,9 +10,9 @@ public class ExtendedPointMatch extends PointMatch
 	 */
 	private static final long serialVersionUID = 1L;
 	
-	final float radius1L, radius1W, radius2L, radius2W;
+	final double radius1L, radius1W, radius2L, radius2W;
 	
-	public ExtendedPointMatch( final Point p1, final Point p2, final float radius1L, final float radius1W, final float radius2L, final float radius2W ) 
+	public ExtendedPointMatch( final Point p1, final Point p2, final double radius1L, final double radius1W, final double radius2L, final double radius2W ) 
 	{
 		super( p1, p2 );
